🔮refactor[#195] 토큰, 인증 역할 분리
